Worked on PCB and schematic
removed intermediary mosfets in the output stage, they take to much spac with the coolers added. it wont fit like that and it should still work without them ontop it is more efficient. Added cooler to PCB, PCB needs to be enlarged to about 135mm and component placing needs to be reconsiderd.